Subtraction class

+++ src/main/java/com/qa/calculator/Calculator.java
@@ -18,19 +18,19 @@ public class Calculator {
 		}
 	}
 
-	public float getNum1() {
+	public double getNum1() {
 		return num1;
 	}
 
-	public void setNum1(float num1) {
+	public void setNum1(double num1) {
 		this.num1 = num1;
 	}
 
-	public float getNum2() {
+	public double getNum2() {
 		return num2;
 	}
 
-	public void setNum2(float num2) {
+	public void setNum2(double num2) {
 		this.num2 = num2;
 	}
